Debug output and some subfield checking
Now reports subfields which are not defined by the format.

For Koha, you probably want to add --ignore=xxx9 as parameter, if
you have any authorised fields, as Koha saves the authority record
number in the 9 subfield.

Removed the XML reading script, run marc_warnings with --debug to
get the same output.